Extract function for extracting secrets from a header to a separate class.
