+++ src/test/java/uk/gov/hmcts/reform/draftstore/service/secrets/SecretsBuilderTest.java
@@ -7,22 +7,22 @@ import static java.util.Collections.nCopies;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class SecretsTest {
+public class SecretsBuilderTest {
 
     @Test
     public void should_validate_number_of_secrets() throws Exception {
         String header =
             String.join(
-                Secrets.SEPARATOR,
+                SecretsBuilder.SEPARATOR,
                 nCopies(3, secretOfLength(Secrets.MIN_SECRET_LENGTH))
             );
 
-        assertThatThrownBy(() -> Secrets.fromHeader(header)).isInstanceOf(SecretsException.class);
+        assertThatThrownBy(() -> SecretsBuilder.fromHeader(header)).isInstanceOf(SecretsException.class);
     }
 
     @Test
     public void should_handle_null_header() throws Exception {
-        Secrets secrets = Secrets.fromHeader(null);
+        Secrets secrets = SecretsBuilder.fromHeader(null);
 
         assertThat(secrets.primary).isNull();
         assertThat(secrets.secondary).isNull();
@@ -34,7 +34,7 @@ public class SecretsTest {
             String ok = secretOfLength(Secrets.MIN_SECRET_LENGTH);
             String tooShort = secretOfLength(Secrets.MIN_SECRET_LENGTH - 1);
 
-            Secrets.fromHeader(ok + Secrets.SEPARATOR + tooShort);
+            SecretsBuilder.fromHeader(ok + SecretsBuilder.SEPARATOR + tooShort);
         }).isInstanceOf(SecretsException.class);
     }
 
@@ -43,7 +43,7 @@ public class SecretsTest {
         String s1 = Strings.repeat("a", Secrets.MIN_SECRET_LENGTH);
         String s2 = Strings.repeat("b", Secrets.MIN_SECRET_LENGTH);
 
-        Secrets secrets = Secrets.fromHeader(s1 + Secrets.SEPARATOR + s2);
+        Secrets secrets = SecretsBuilder.fromHeader(s1 + SecretsBuilder.SEPARATOR + s2);
 
         assertThat(secrets.primary).isEqualTo(s1);
         assertThat(secrets.secondary).isEqualTo(s2);
@@ -53,7 +53,7 @@ public class SecretsTest {
     public void should_convert_header_with_one_secret_to_object() throws Exception {
         String secret = Strings.repeat("x", Secrets.MIN_SECRET_LENGTH);
 
-        Secrets secrets = Secrets.fromHeader(secret);
+        Secrets secrets = SecretsBuilder.fromHeader(secret);
 
         assertThat(secrets.primary).isEqualTo(secret);
         assertThat(secrets.secondary).isNull();
